afxdp/worker/cos: relocate inline tests to cos_tests.rs
Mirrors the umem/{mod,tests}.rs and forwarding/{mod,tests}.rs split
patterns. worker/cos.rs reaches 2,689 LOC with a 1,914-LOC
`#[cfg(test)] mod tests { ... }` block at the end (file becomes mostly
tests).

Mechanical changes:
- Inline `#[cfg(test)] mod tests { ... }` block (lines 775-2688)
  extracted to sibling cos_tests.rs (1,917 lines incl 5-line header)
- 4-space dedent applied uniformly
- mod.rs (cos.rs) ends with:
    #[cfg(test)]
    #[path = "cos_tests.rs"]
    mod tests;
- worker/ is a directory module but cos.rs is a flat file (not its own
  directory), so the explicit `#[path]` attribute matches the
  mpsc_inbox.rs/test_fixtures.rs convention used elsewhere
- pre-rustfmt'd cos_tests.rs to match prior PRs' Copilot expectations

cos.rs drops from 2,689 → 778 LOC.

cargo build --release: clean. cargo test --release: 878 passed.
